fix: shutdown handler potential fix

+++ ROS2/colcon_ws/src/beaumont_pkg/beaumont_pkg/shutdown_handler.py
@@ -2,7 +2,7 @@ import os
 import rclpy
 from rclpy.node import Node
 
-from std_msgs.msg import String
+from std_msgs.msg import Bool
 
 
 class ShutdownHandler(Node):
@@ -10,7 +10,7 @@ class ShutdownHandler(Node):
     def __init__(self):
         super().__init__('shutdown_handler')
         self.subscription = self.create_subscription(
-            bool,
+            Bool,
             'shutdown',
             self.listener_callback,
             10)
@@ -18,11 +18,12 @@ class ShutdownHandler(Node):
 
     def listener_callback(self, msg):
         #TODO: figure out which ip is which
-        os.exec("ssh pizero0@192.168.2.13 'poweroff'")
-        os.exec("ssh pizero1@192.168.2.10 'poweroff'")
-        os.exec("ssh pizero2@192.168.2.11 'poweroff'")
-        os.exec("ssh pizero3@192.168.2.12 'poweroff'")
-        os.exec("poweroff")
+        if msg.data == True:
+            os.exec("ssh pizero0@192.168.2.13 'poweroff'")
+            os.exec("ssh pizero1@192.168.2.10 'poweroff'")
+            os.exec("ssh pizero2@192.168.2.11 'poweroff'")
+            os.exec("ssh pizero3@192.168.2.12 'poweroff'")
+            os.exec("poweroff")
